Sidebar menu functions, work on getting it persistent
