+++ app/src/main/AndroidManifest.xml
@@ -34,8 +34,7 @@
 
         <activity
             android:name=".login.LoginView"
-            android:label="@string/title_activity_login"
-            android:theme="@style/AppTheme.NoActionBar">
+            android:label="@string/title_activity_login">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
